\name{integral}
\alias{integral}
\title{integral(x, y)}
\usage{
integral(x, y, maxParticiones = TRUE, Particiones = 3)
}
\description{
definite integral calculated by means of the data x and y
}
\examples{
integral(x, y, maxParticiones = TRUE, Particiones = 3)
}
\value{
integral defined between x [1] and the value x [length(x)]
}
\arguments{
\item{x}{
List of values of x in a function
}
\item{y}{
List of values of y in a function
}
\item{maxParticiones}{
want to calculate the maximum of partitions?
}
\item{particiones}{
how many partitions should I do if I do not want to calculate the maximum possible
}
}
\author{Diego Barajas "barajas-d@javeriana.edu.co and Brandonn Cruz "Brandonn_Cruz@javeriana.edu.co"}

